refactor Profile View
